Make the .manifest section fixed length
This avoids the hack from PR #106 with a less ugly one: an empty (all zeros)
.manifest section is added before computing the Merkle roots, then it is
updated with the correct serialized Manifest, padding so that the length remains
unchanged at 4 kb. In this way, the Merkle roots are more robust as they are not
affected by the length of the serialized Manifest.