Decouple ecalls from the trait
We'd like to add a fork of getrandom (similar to the existing ones)
which uses ecall to generate randomness. However, depending on SDK has a
problem: on non-riscv platforms the SDK indirectly depends on
`getrandom`. So this is not really a cycle but cargo thinks it is, so we
have to break it anyway. We will do it by separating the ecalls into a
crate. So far they were implemented as trait methods directly, but we
don't want this trait long-term, so trying to put it into a crate seems
wasted.

This commit thus decouples the calls from the trait impl by implementing
them as standalone functions and then simply calls them in the trait
impl. Also, because we want to make all calls public, this simplifies
the macros to just make all of them public.