Make ecalls `unsafe`
These functions take pointers which they dereference which allows safe
code to cause UB. Thus they need to be `unsafe`. Ideally, safety would
be enforced by the API which should be done eventually, this is just a
step forward towards correctness.

Note that we don't make the trait methods `unsafe` to keep the commit
simple.