Rename back the bitcoin libraries
The renamed libraries were annoying to deal with in the consumers if the
consumers needed to use other libraries that are already using the
bitcoin libraries. Renaming them back allows using `[patch.crates-io.*]`
in the application root to override all of them at once. This is also
done for the `bitcoin` app to demonstrate it works.

This also renames `bitcoin-hashes` back to `bitcoin_hashes` to avoid
confusion.