Support auxiliary entropy
Generally, applications tend to use auxiliary entropy but this wasn't
implemented yet. Strictly speaking, `_no_aux_rand` did it implicitly but
that was wrong and would require awkward `cfg` flags.

This commit adds another parameter to provide entropy. When set to null
the code behaves like before and when set to non-null it uses the
entropy to protect the signature against side channel attacks.

Notably, `no_aux_rand` is still wrong for simplicity of the commit. This
needs to be fixed later.