Refactor the VectorAccumulator to allow streaming verification.
As we might not receive the entire Merkle proofs in a single
message, checking them in a 'streaming' mode allows to save both
memory and unnecessary copying/allocations.

A new StreamingVectorAccumulator trait (and the companions
InclusionProofVerifier and UpdateProofVerifier) implements the
interface for verifying the inclusion and update proofs in
streaming fashion.

By splitting the verification logic out of the VectorAccumulator
trait into a new VectorAccumulatorVerifier trait, we can provide a
blanket implementation for any struct that implements the new
StreamingVectorAccumulator, while keeping the same interface for
the non-streaming version of the methods.