Add ECALL and minimal app-sdk functionality for SLIP-21
The ECALL gets the full derivation, leaving the possibility in the
future of adding per-app path-restrictions in the V-App's manifest.

This introduces a custom derivation scheme for the SLIP21 that uses
a different master seed (in turn derived from the standard SLIP21
hierarchy). This allows to stay compatible with Bolos, whose
SLIP21 implementation is limited and doesn't currently allow
Vanadium to access arbitrary SLIP-21 derivations.