Custom send/receive flow that does not process unnecessary events
Comm::reply calls Comm::apdu_send, which receives an event before
sending the APDU. This is problematic when there is no even, as it
ends up waiting for the next ticker event which takes up to 0.1s.

In io_exchange, we want to send an APDU and receive the response
as fast as possible, and do not expect processing anything else.