Add a ResettableHasher trait.
In some contexts, many independent hashes have to be computed. If
the hasher has the capability of resetting its context, this can
be used to avoid allocating a new hasher for each hash.

The ResettableHasher trait extends the Hasher trait by providing
this capability.

This commit also restricts the MerkleAccumulator to use a
ResettableHasher, although it does not yet take advantage of it.