Simplified protocol for sending a buffer and sending a panic
Both are doing exactly the same thing, so the protocol was changed
by adding a single-byte label in the first message (specifying the
kind of buffer); the message for the subsequent chunks (in case
the message doesn't fit in the first chunk) is now shared.

This makes it more generalizable, and it also saves a few bytes
for each subsequent message.